requirements.txt
openai
tiktoken
langchain
langchain-core
langchain-community
langchainhub 
faiss-cpu 
unstructured